--- a/random/random.go
+++ b/random/random.go
@@ -3,13 +3,29 @@ package random
 import (
 	"math/rand"
 	"strings"
+	"sync"
 	"time"
 )
 
 type (
 	Random struct {
+		lock sync.Mutex
+		src  rand.Source
 	}
 )
+
+func (r *Random) Int63() (n int64) {
+	r.lock.Lock()
+	n = r.src.Int63()
+	r.lock.Unlock()
+	return
+}
+
+func (r *Random) Seed(seed int64) {
+	r.lock.Lock()
+	r.src.Seed(seed)
+	r.lock.Unlock()
+}
 
 // Charsets
 const (
@@ -27,8 +43,8 @@ var (
 )
 
 func New() *Random {
-	rand.Seed(time.Now().UnixNano())
-	return new(Random)
+	src := rand.NewSource(time.Now().UnixNano())
+	return &Random{src: src}
 }
 
 func (r *Random) String(length uint8, charsets ...string) string {
@@ -38,7 +54,7 @@ func (r *Random) String(length uint8, charsets ...string) string {
 	}
 	b := make([]byte, length)
 	for i := range b {
-		b[i] = charset[rand.Int63()%int64(len(charset))]
+		b[i] = charset[r.Int63()%int64(len(charset))]
 	}
 	return string(b)
 }