--- a/random/random.go
+++ b/random/random.go
@@ -1,31 +1,18 @@
 package random
 
 import (
-	"math/rand"
+	"bufio"
+	"crypto/rand"
+	"io"
 	"strings"
 	"sync"
-	"time"
 )
 
 type (
 	Random struct {
-		lock sync.Mutex
-		src  rand.Source
+		readerPool sync.Pool
 	}
 )
-
-func (r *Random) Int63() (n int64) {
-	r.lock.Lock()
-	n = r.src.Int63()
-	r.lock.Unlock()
-	return
-}
-
-func (r *Random) Seed(seed int64) {
-	r.lock.Lock()
-	r.src.Seed(seed)
-	r.lock.Unlock()
-}
 
 // Charsets
 const (
@@ -43,8 +30,10 @@ var (
 )
 
 func New() *Random {
-	src := rand.NewSource(time.Now().UnixNano())
-	return &Random{src: src}
+	p := sync.Pool{New: func() interface{} {
+		return bufio.NewReader(rand.Reader)
+	}}
+	return &Random{readerPool: p}
 }
 
 func (r *Random) String(length uint8, charsets ...string) string {
@@ -52,11 +41,43 @@ func (r *Random) String(length uint8, charsets ...string) string {
 	if charset == "" {
 		charset = Alphanumeric
 	}
-	b := make([]byte, length)
-	for i := range b {
-		b[i] = charset[r.Int63()%int64(len(charset))]
+
+	charsetLen := len(charset)
+	if charsetLen > 255 {
+		charsetLen = 255
 	}
-	return string(b)
+	maxByte := 255 - (256 % charsetLen)
+
+	reader := r.readerPool.Get().(*bufio.Reader)
+	defer r.readerPool.Put(reader)
+
+	b := make([]byte, length)
+	rs := make([]byte, length+(length/4)) // perf: avoid read from rand.Reader many times
+	var i uint8 = 0
+
+	// security note:
+	// we can't just simply do b[i]=randomStringCharset[rb%len(randomStringCharset)],
+	// for example, when len(charsets) is 52, and rb is [0, 255], 256 = 52 * 4 + 48.
+	// this will make the first 48 characters more possibly to be generated then others.
+	// so we have to skip bytes when rb > maxByte
+
+	for {
+		_, err := io.ReadFull(reader, rs)
+		if err != nil {
+			panic("unexpected error happened when reading from bufio.NewReader(crypto/rand.Reader)")
+		}
+		for _, rb := range rs {
+			if rb > byte(maxByte) {
+				// Skip this number to avoid bias.
+				continue
+			}
+			b[i] = charset[rb%byte(charsetLen)]
+			i++
+			if i == length {
+				return string(b)
+			}
+		}
+	}
 }
 
 func String(length uint8, charsets ...string) string {