--- a/random/random.go
+++ b/random/random.go
@@ -30,6 +30,7 @@ var (
 )
 
 func New() *Random {
+	// https://tip.golang.org/doc/go1.19#:~:text=Read%20no%20longer%20buffers%20random%20data%20obtained%20from%20the%20operating%20system%20between%20calls
 	p := sync.Pool{New: func() interface{} {
 		return bufio.NewReader(rand.Reader)
 	}}