@ichigozero 2023-08-12
> maybe https://github.com/labstack/echo/blob/e6b96f8873fed46e71e0d34cddb81c533167f954/middleware/util.go#L69 would be suitable here also.

Thank you for the comment. I will take a look the link you mentioned and update the PR accordingly.